+++ src/frontends/lean/definition_cmds.h
@@ -9,7 +9,6 @@ Author: Leonardo de Moura
 #include "frontends/lean/decl_attributes.h"
 #include "frontends/lean/decl_util.h"
 namespace lean {
-
 environment definition_cmd_core(parser & p, decl_cmd_kind k, cmd_meta const & meta);
 
 environment ensure_decl_namespaces(environment const & env, name const & full_n);