+++ src/frontends/lean/token_table.cpp
@@ -101,7 +101,7 @@ void init_token_table(token_table & t) {
 
     char const * commands[] =
         {"theorem", "axiom", "axioms", "variable", "protected", "private", "hide",
-         "definition", "meta", "mutual", "example", "noncomputable",
+         "definition", "meta", "mutual", "example", "noncomputable", "abbreviation",
          "variables", "parameter", "parameters", "constant", "constants",
          "using_well_founded", "[whnf]",
          "end", "namespace", "section", "prelude",